* FIX [mqtt_decode] fix MQTT v5 protocol violations

+++ src/supplemental/mqtt/mqtt_public.c
@@ -957,7 +957,7 @@ nng_mqtt_unsubscribe_async(nng_mqtt_client *client, nng_mqtt_topic *sbs, size_t 
 }
 // send a blocking sub msg until suback returns
 int
-nng_mqtt_subscribe(nng_socket sock, nng_mqtt_topic_qos *sbs, size_t count, property *pl)
+nng_mqtt_subscribe(nng_socket sock, nng_mqtt_topic_qos *sbs, uint32_t count, property *pl)
 {
 	int rv = 0;
 	// create a SUBSCRIBE message
@@ -988,7 +988,7 @@ nng_mqtt_subscribe(nng_socket sock, nng_mqtt_topic_qos *sbs, size_t count, prope
 	if (msg) {
 		uint8_t *code = nng_mqtt_msg_get_suback_return_codes(msg, &count);
 		log_info("bridge: subscribe aio result %d", rv);
-		for (int i=0; i<count; ++i) {
+		for (uint32_t i=0; i<count; ++i) {
 			log_info("bridge: suback code %d ", *(code + i));
 		}
 		nng_msg_free(msg);
